added getTokensFromCollection in utils
